ds_requirements.txt
deepface==0.0.53
Keras==2.4.0
Keras-Preprocessing==1.1.2
mtcnn==0.1.0
numpy==1.19.5
opencv-python==4.5.2.52
Pillow==8.2.0
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
retina-face==0.0.4
scipy==1.6.3
six==1.15.0
sqlparse==0.4.1
tensorboard==2.4.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.0
tensorflow-estimator==2.4.0
tqdm==4.61.0
urllib3==1.26.5
yolo-v4==0.5

Django==3.0.2
django-celery-results==2.0.1
django-redis==4.12.1
djangorestframework==3.12.4
